multimodel_rag.main-main/multi-media-rag1-main/requirements.txt
streamlit>=1.39.0
faiss-cpu>=1.8.0
scikit-learn>=1.5.0
groq>=0.11.0
pdfplumber>=0.11.0
numpy>=1.26.0
